+++ app/components/BrowseProducts.jsx
@@ -1,8 +1,9 @@
 import React, {Component} from 'react';
-
+import {connect} from 'react-redux';
+import {loadProducts} from 'APP/app/reducers/products';
 import {Link} from 'react-router';
 
-export default class BrowseProducts extends Component {
+export class BrowseProducts extends Component {
 
   render () {
     return (
@@ -10,11 +11,15 @@ export default class BrowseProducts extends Component {
         {
           this.props.products.map(product => {
             return (
-              <ul key={product.id} className="list-unstyled">
-                <li>
-                  <Link to={'/products/' + product.id}> { product.name }</Link>
-                </li>
-              </ul>
+              <div key={product.id} className="list-unstyled col-xs-3 product-card">
+                  <Link to={'/products/' + product.id}>
+                  <img src={product.image} />
+                  <h3>{ product.name }</h3>
+                  <p>{ product.category }</p>
+                  <p>{ product.description }</p>
+                  <p>{ product.name }</p>
+                  </Link>
+              </div>
             );
           })
         }
@@ -22,5 +27,25 @@ export default class BrowseProducts extends Component {
     );
   }
 }
+
+const mapStateToProps = state => {
+  return {
+    products: state.products
+  };
+};
+
+const mapDispatchToProps = dispatch => {
+  return {
+    onLoadProducts: function () {
+      const thunk = loadProducts();
+      dispatch(thunk);
+    }
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(BrowseProducts);
