+++ app/main.jsx
@@ -12,9 +12,8 @@ import WhoAmI from './components/WhoAmI';
 import Checkout from './components/Checkout';
 import AddProduct from './components/AddProduct';
 import Orders from './components/Orders';
-
-import ProductDetailContainer from './components/ProductDetailContainer';
-import BrowseProductsContainer from './components/BrowseProductsContainer';
+import ProductDetail from './components/ProductDetail';
+import BrowseProducts from './components/BrowseProducts';
 
 import {loadProducts, setSelectedProductId} from './reducers/products';
 
@@ -49,8 +48,8 @@ render(
         <Route path="/checkout" component={Checkout} />
         <Route path="/products/add" component={AddProduct} />
         <Route path="/orders" component={Orders} />
-        <Route path="/products" component={BrowseProductsContainer} onEnter={onBrowse} />
-        <Route path="/products/:id" component={ProductDetailContainer} onEnter={setProduct} />
+        <Route path="/products" component={BrowseProducts} onEnter={onBrowse} />
+        <Route path="/products/:id" component={ProductDetail} onEnter={setProduct} />
       </Route>
     </Router>
   </Provider>,
