+++ public/stylesheets/products.scss
@@ -13,3 +13,4 @@
   margin: 5px;
   padding-top: 15px;
 }
+
