+++ public/stylesheets/style.css
@@ -1,2 +1,14 @@
-body {
-  background: #ffffff; }
+.products-container {
+  max-width: 800px;
+  border-left: 5px;
+  border-right: 5px; }
+
+.product-card {
+  border: 2px solid #595959;
+  border-radius: 10px;
+  box-shadow: 3px 3px 3px grey;
+  background-color: #ffffff;
+  clear: all;
+  height: 350px;
+  margin: 5px;
+  padding-top: 15px; }
